backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
